refactor: divide
